requirements.txt
streamlit==1.41.1
seaborn==0.13.2
matplotlib
scikit-learn 
textblob
scikit-surprise
ensure
python-box
pyyaml
fastapi
uvicorn[standard]